requirements.txt
Django==5.0.6
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2==2.9.9
redis==5.0.1
celery==5.3.4
Pillow==10.1.0
requests==2.31.0
stripe==7.4.0
twilio==8.10.0
django-cors-headers==4.3.1
django-filter==23.5
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
PyJWT==2.8.1
cryptography==41.0.7
drf-yasg==1.21.7
django-extensions==3.2.3
python-dateutil==2.8.2
pytz==2023.3
